240428 점수 기록 DB 설계 및 추가 [ for PointCard ] , 격언 문구 DB 설계 및 추가 [ NavStatisticsPage ], Toit 점수 계산 로직 추가 및 점수 적용 [ProfilePage, TotalSummaryUnit]

+++ app/src/main/java/com/one/toit/ui/compose/ui/unit/ToitPointCard.kt
@@ -93,7 +93,7 @@ import kotlin.math.roundToInt
 @Composable
 fun ToitPointCard(
     modifier: Modifier = Modifier,
-    toitPoint:Int,
+    toitPoint:Long,
     durationMillis:Int = 1500
 ){
